Use globally-unique binding IDs for syntax-rules literal identity (#1272) (#1284)
Replace per-frame slot numbers (u16) with monotonic binding IDs (u32)
when comparing literal identities in syntax-rules. Slot indices are
scoped to their owning Compiler frame, so two different bindings in
nested lambdas occupying the same register falsely compared as equal.

+++ src/types.zig
@@ -377,7 +377,7 @@ pub const Transformer = struct {
     def_env: ?*std.StringHashMap(Value) = null,
     def_env_val: Value = NIL,
     custom_ellipsis: ?[]const u8 = null,
-    literal_bound: []u16 = &.{},
+    literal_bound: []u32 = &.{},
     let_syntax_peer_names: [][]const u8 = &.{},
     let_syntax_peer_vals: []Value = &.{},
 };
